[#153] Use same common component for posix and process_local
